+++ __mocks__/pwaRegisterMock.js
@@ -1,0 +1,7 @@
+export function useRegisterSW() {
+  return {
+    offlineReady: [false, () => {}],
+    needRefresh: [false, () => {}],
+    updateServiceWorker: () => {},
+  };
+}